Gradually chopping down the compile errors on the refactor

+++ Gravity.Server/Pipeline/OutgoingMessage.cs
@@ -6,7 +6,7 @@ namespace Gravity.Server.Pipeline
 {
     internal class OutgoingMessage : Message, IOutgoingMessage
     {
-        public uint StatusCode { get; set; }
+        public ushort StatusCode { get; set; }
         public string ReasonPhrase { get; set; }
     }
 }
